\name{show_foragers}
\alias{show_foragers}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Show forager skill curves from sim_foragers}
\description{
  Plots skill curves from simulated foragers generated by sim_foragers
}
\usage{
show_foragers( sim , cols=c(rangi2,"red","black","green") )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
    \item{sim}{Output of sim_foragers}
    \item{cols}{Colors for plotting}
}
\details{
  This function
}
\references{}
\seealso{ \code{sim_foragers} }
\examples{

dat <- sim_foragers( N_foragers=1 , prob_drop_fail=0 , prob_pool=0 , prob_assistant=0.5 , coop=c(0,0,0,0,0,1) , a=c(3,log(10),0.5) )
show_foragers(dat)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }

